finish navigation at all
