+++ frontend/src/App.module.scss
@@ -14,5 +14,5 @@ html{
 
 body{
     background-color: $primary-color;
-    color: white;
+    color: $primary-font-color;
 }
