+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.module.scss'
 
 import Home from './pages/home/Home'
-import Navigation from './components/navigation/Navigation'
+import Navigation from './components/layout/navigation/Navigation'
 import Category from './pages/category/Category'
 import { Routes, Route, BrowserRouter } from 'react-router-dom'
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client'
