+++ frontend/src/assets/_mixin.module.scss
@@ -1,0 +1,12 @@
+@mixin item {
+	border-radius: 20px;
+	color: $primary-font-color;
+	transition: all 0.6s;
+	cursor: pointer;
+
+	&:hover {
+		background-color: $primary-color;
+		color: $secondary-font-color;
+		box-shadow: 2px 2px 10px $primary-color;
+	}
+}
