+++ frontend/src/components/layout/navigation/nav-items/NavItems.module.scss
@@ -1,4 +1,5 @@
-@import './../../../assets/colors.module';
+@import './../../../../assets/colors.module';
+@import './../../../../assets/mixin.module';
 
 .items {
 	display: flex;
@@ -24,24 +25,15 @@
 		}
 	}
 	&__item {
+		@include item;
 		margin: 1em 0;
-		cursor: pointer;
 		padding: 1em;
-		border-radius: 20px;
-		transition: all 0.6s;
-		color: $primary-font-color;
 		@media (min-width: 768px) {
 			padding: 0.5em 1em;
 			margin: 0 1em;
 		}
 		&-icon {
 			margin-right: 0.4em;
-		}
-
-		&:hover {
-			background-color: $primary-color;
-			color: $secondary-font-color;
-			box-shadow: 2px 2px 10px $primary-color;
 		}
 	}
 	&__link {
