+++ frontend/src/components/layout/navigation/nav-items/NavItems.tsx
@@ -1,14 +1,23 @@
-import styles from './NavItems.module.scss'
+import { useCategory } from '../../../../hooks/useCategory'
 import { Link } from 'react-router-dom'
-import { useCategory } from '../../../hooks/useCategory'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faGamepad, faTag } from '@fortawesome/free-solid-svg-icons'
-import { useState, useEffect } from 'react'
+
+import styles from './NavItems.module.scss'
+
+interface Props {
+	matches: boolean
+}
+
+interface Category {
+	id: string
+	attributes: any
+}
 
 const gamepadIcon = <FontAwesomeIcon icon={faGamepad} className={styles['items__item-icon']} />
 const tagIcon = <FontAwesomeIcon icon={faTag} className={styles['items__item-icon']} />
 
-const NavItems = props => {
+const NavItems = (props: Props) => {
 	const { loading, error, data } = useCategory()
 
 	if (loading) return <p>loading...</p>
@@ -17,14 +26,14 @@ const NavItems = props => {
 	return (
 		<div className={styles.items}>
 			<ul className={styles['items__list']}>
-				{data.categories.data.map(category => (
-					<Link to={`/category/${category.id}`} className={styles['items__link']}>
+				{data.categories.data.map((category: Category) => (
+					<Link to={`/category/${category.id}`} className={styles['items__link']} key={category.id}>
 						<li key={category.id} className={styles['items__item']}>
 							{category.attributes.name}
 						</li>
 					</Link>
 				))}
-				{!props.matchesMD && (
+				{!props.matches && (
 					<div>
 						<hr />
 						<li className={styles['items__item']}>{gamepadIcon}Biblioteka Gier</li>
