+++ frontend/src/components/layout/navigation/nav-sidebar/NavSidebar.tsx
@@ -1,6 +1,7 @@
-import styles from './NavSidebar.module.scss'
-import Social from '../../social/Social'
 import NavItems from '../nav-items/NavItems'
+import Social from '../../../social/Social'
+
+import styles from './NavSidebar.module.scss'
 
 interface Props {
 	mobileNav: boolean
@@ -9,8 +10,8 @@ interface Props {
 const NavSidebar = (props: Props) => {
 	return (
 		<div className={`${styles.sidebar} ${props.mobileNav ? styles.active : ''}`}>
-			<NavItems />
-			<Social/>
+			<NavItems matches={false} />
+			<Social />
 		</div>
 	)
 }
