--- frontend/src/components/navigation/Navigation.module.scss
@@ -1,24 +1,0 @@
-@import './../../assets/colors.module';
-
-.nav{
-    position: fixed;
-    left: 0;
-    right: 0;
-    top: 0;
-    background-color: $secondary-color;
-    color: $secondary-font-color;
-    z-index: 1000;
-    &__container{
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin: 0 auto;
-        padding:0 1em ;
-        width: 100%;
-        max-width: 1200px;
-    }
-    &__social{
-        font-size: 1.6rem;
-    }
-}
-
