--- frontend/src/components/navigation/nav-bar/NavBar.module.scss
@@ -1,24 +1,0 @@
-@import './../../../assets/colors.module';
-
-.header {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-    padding: .7em;
-	&__container{
-		display: flex;
-		justify-content: center;
-		align-items: center;
-	}
-	&__logo {
-		width: 35px;
-        margin-right: .5em;
-	}
-	&__heading {
-        font-size: 1.4rem;
-		margin-right: 1em;
-		color: $secondary-font-color;
-        
-
-	}
-}
